Refactored PygamUI and Class DayScreen is now StoryScreen
